Enhance fix from previous commit
The original fix would only put those items on the pasteboard that currently have an item cell representing them. This resulted in a sort of arbitrary selection for large selection sets (think hundreds or thousands of items via CMD-A).

Now, not all of the dragging items are provided with a dragging image. The maximum number of draggable items is also limited to 2000 - anything more would probably not amuse our user anyhow.